Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -66,6 +66,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){ 
+            abort(403, 'Accesss denied'); 
+        }
         return view('questions.edit', compact('question')); 
     }
 
@@ -78,6 +81,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){ 
+            abort(403, 'Accesss denied'); 
+        }
+
         $question->update($request->only('title', 'body')); 
 
         return redirect()->route('questions.index')->with('success', 'Your question has been updated!');  
@@ -91,6 +98,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){ 
+            abort(403, 'Accesss denied'); 
+        }
+
         $question->delete(); 
         
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted!');
